requirements.txt
streamlit
weaviate-client
sentence-transformers
pymongo